requirements.txt
Flask~=3.1.2
pillow~=11.0.0
Werkzeug~=3.1.3